Added 3 files
allwords.py: creates a list of unique words that could be used to aid in translating my resume.
jobApplicationQandA: compiling a list of answers to questions that might be useful in future job applications
references.csv: maintaining a permanent list of references so hopefully I don't need to ask people in the future.